requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
# Для подключения к PostgreSQL, даже если мы его не используем в этом простом приложении
psycopg2-binary==2.9.9